cartrack/scraper/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
playwright==1.44.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
python-dotenv==1.0.1
httpx==0.27.0
apscheduler==3.10.4
pydantic==2.7.1
pydantic-settings==2.3.1
beautifulsoup4==4.12.3
lxml==5.2.2
python-multipart==0.0.9
plyer==2.1.0
